# Reduce logs in CLI scripts and plugin loader